+++ tests/esg.rs
@@ -2,5 +2,9 @@ mod common;
 
 #[path = "esg/live.rs"]
 mod esg_live;
+#[path = "esg/live_involvement.rs"]
+mod esg_live_involvement;
 #[path = "esg/offline.rs"]
 mod esg_offline;
+#[path = "esg/offline_involvement.rs"]
+mod esg_offline_involvement;